requirements.txt
openai>=1.0.0
pinecone>=5.0.0
PyPDF2>=3.0.1
rank-bm25>=0.2.2
python-dotenv>=1.0.0
numpy>=1.26.0
streamlit>=1.38.0
pandas>=2.2.0
pycryptodome>=3.20.0


